Feature/views

+++ db/migrate/20200817220146_create_weights.rb
@@ -1,7 +1,7 @@
 class CreateWeights < ActiveRecord::Migration[6.0]
   def change
     create_table :weights do |t|
-      t.integer    :value,            null: false
+      t.float      :value,            null: false
       t.integer    :unity,            null: false
       t.date       :registration_date
       t.references :user,             null: false, foreign_key: true
